move custom_auth
